+++ CHANGELOG.md
@@ -3,11 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.7.2
+## 3.4.7.3
 
 ### New Features
 
 ### Fixes
-- Launching an instance with a non valid account set causing it to not launch
 
 ### Misc